Add correction 2.
In Problem 2(e), compute $H(x)$ from $p=0.003$ from (c).  The reference to part (b) is incorrect.  (b) is a plot that
does not specify $p$ and the value from (a) is unrelated to the $p$ value used to
generate the input file. 2(e) was updated to reflect this change.